requirements.txt
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.8
imbalanced-learn==0.13.0
imblearn==0.0
joblib==1.4.2
kiwisolver==1.4.8
llvmlite==0.44.0
matplotlib==3.10.0
numba==0.61.0
numpy==2.1.3
nvidia-nccl-cu12==2.25.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-learn==1.5.2
scipy==1.15.1
seaborn==0.13.2
shap==0.46.0
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
threadpoolctl==3.5.0
tqdm==4.67.1
tzdata==2025.1
xgboost==2.1.3
